backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.4.0
aiohttp>=3.9.0
lxml>=4.9.0
feedparser>=6.0.0
python-dateutil>=2.8.0
